[Design] : FindingTheCharacter GifView Preview 생성
Git : #8
