+++ Projects/Share/AD_Utils/Resources/Previews/Gifs.xcassets/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
