+++ Projects/Share/AD_Utils/Resources/Previews/Gifs.xcassets/FindingTheCharacter_Preview2.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "data" : [
+    {
+      "filename" : "FindingTheCharacter_Preview2.gif",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
